+++ src/dune/promotion.ml
@@ -1,31 +1,64 @@
 open! Stdune
 
+let staging_area = Path.Build.relative Path.Build.root ".promotion-staging"
+
 module File = struct
   type t =
     { src : Path.Build.t
+    ; staging : Path.Build.t option
     ; dst : Path.Source.t
     }
 
-  let to_dyn { src; dst } =
+  let in_staging_area source = Path.Build.append_source staging_area source
+
+  let to_dyn { src; staging; dst } =
     let open Dyn.Encoder in
-    record [ ("src", Path.Build.to_dyn src); ("dst", Path.Source.to_dyn dst) ]
+    record
+      [ ("src", Path.Build.to_dyn src)
+      ; ("staging", option Path.Build.to_dyn staging)
+      ; ("dst", Path.Source.to_dyn dst)
+      ]
 
   let db : t list ref = ref []
 
-  let register t = db := t :: !db
+  let register_dep ~source_file ~correction_file =
+    db :=
+      { src = snd (Path.Build.split_sandbox_root correction_file)
+      ; staging = None
+      ; dst = source_file
+      }
+      :: !db
 
-  let promote { src; dst } =
-    let src_exists = Path.exists (Path.build src) in
+  let register_intermediate ~source_file ~correction_file =
+    let staging = in_staging_area source_file in
+    Path.mkdir_p (Path.build (Option.value_exn (Path.Build.parent staging)));
+    Unix.rename
+      (Path.Build.to_string correction_file)
+      (Path.Build.to_string staging);
+    let src = snd (Path.Build.split_sandbox_root correction_file) in
+    db := { src; staging = Some staging; dst = source_file } :: !db
+
+  let promote { src; staging; dst } =
+    let correction_file = Option.value staging ~default:src in
+    let correction_exists = Path.exists (Path.build correction_file) in
     Console.print
-      (Format.sprintf
-         ( if src_exists then
-           "Promoting %s to %s.@."
-         else
-           "Skipping promotion of %s to %s as the file is missing.@." )
-         (Path.to_string_maybe_quoted (Path.build src))
-         (Path.Source.to_string_maybe_quoted dst));
-    if src_exists then
-      Io.copy_file ~src:(Path.build src) ~dst:(Path.source dst) ()
+      ( if correction_exists then
+        Format.sprintf "Promoting %s to %s.@."
+          (Path.to_string_maybe_quoted (Path.build src))
+          (Path.Source.to_string_maybe_quoted dst)
+      else
+        (Format.sprintf
+           "Skipping promotion of %s to %s as the %s is missing.@.")
+          (Path.to_string_maybe_quoted (Path.build src))
+          (Path.Source.to_string_maybe_quoted dst)
+          ( match staging with
+          | None ->
+              "file"
+          | Some staging ->
+              Format.sprintf "staging file (%s)"
+                (Path.to_string_maybe_quoted (Path.build staging)) ) );
+    if correction_exists then
+      Io.copy_file ~src:(Path.build correction_file) ~dst:(Path.source dst) ()
 end
 
 let clear_cache () = File.db := []
@@ -37,7 +70,7 @@ module P = Persistent.Make (struct
 
   let name = "TO-PROMOTE"
 
-  let version = 1
+  let version = 2
 end)
 
 let db_file = Path.relative Path.build_dir ".to-promote"
@@ -53,10 +86,11 @@ let dump_db db =
 let load_db () = Option.value ~default:[] (P.load db_file)
 
 let group_by_targets db =
-  List.map db ~f:(fun { File.src; dst } -> (dst, src))
+  List.map db ~f:(fun { File.src; staging; dst } -> (dst, (src, staging)))
   |> Path.Source.Map.of_list_multi
   (* Sort the list of possible sources for deterministic behavior *)
-  |> Path.Source.Map.map ~f:(List.sort ~compare:Path.Build.compare)
+  |> Path.Source.Map.map
+       ~f:(List.sort ~compare:(fun (x, _) (y, _) -> Path.Build.compare x y))
 
 type files_to_promote =
   | All
@@ -83,15 +117,18 @@ let do_promote db files_to_promote =
     match srcs with
     | [] ->
         assert false
-    | src :: others ->
+    | (src, staging) :: others ->
         (* We remove the files from the digest cache to force a rehash on the
            next run. We do this because on OSX [mtime] is not precise enough
            and if a file is modified and promoted quickly, it will look like it
-           hasn't changed even though it might have. *)
+           hasn't changed even though it might have.
+
+           aalekseyev: this is probably unnecessary now, depending on when
+           [do_promote] runs (before or after [invalidate_cached_timestamps]) *)
         List.iter dirs_to_clear_from_cache ~f:(fun dir ->
             Cached_digest.remove (Path.append_source dir dst));
-        File.promote { src; dst };
-        List.iter others ~f:(fun path ->
+        File.promote { src; staging; dst };
+        List.iter others ~f:(fun (path, _staging) ->
             Format.eprintf " -> ignored %s.@."
               (Path.to_string_maybe_quoted (Path.build path)))
   in
@@ -113,7 +150,8 @@ let do_promote db files_to_promote =
       in
       Path.Source.Map.to_list by_targets
       |> List.concat_map ~f:(fun (dst, srcs) ->
-             List.map srcs ~f:(fun src -> { File.src; dst }))
+             List.map srcs ~f:(fun (src, staging) ->
+                 { File.src; staging; dst }))
 
 let finalize () =
   let db =